Confirmation for New Game button

+++ app/assets/javascripts/deloitte_game/pages/results.js.coffee
@@ -4,5 +4,9 @@
 DeloitteGame.Pages.Results = 
   init: ->
     $('.new-game-bt').on 'click', ->
-      localStorage.clear()
-      true
+      confirmation = confirm "Are you sure you want to START THE GAME OVER AGAIN?"
+      if confirmation
+        localStorage.clear()
+        true
+      else
+        false
